[test] TripController 테스트 분리 (#361)
